+++ services/marketplace/order/resource.go
@@ -337,6 +337,12 @@ func (r *MarketplaceOrderResource) Schema(ctx context.Context, req resource.Sche
 				Validators: []validator.String{
 					stringvalidator.RegexMatches(regexp.MustCompile(`^-?\d{0,12}(?:\.\d{0,10})?$`), ""),
 				}},
+			"new_plan_billing_mode": schema.StringAttribute{
+				Computed: true,
+				PlanModifiers: []planmodifier.String{
+
+					stringplanmodifier.UseStateForUnknown(),
+				}, MarkdownDescription: "New Plan Billing Mode"},
 			"new_plan_name": schema.StringAttribute{
 				Computed: true,
 				PlanModifiers: []planmodifier.String{
@@ -410,6 +416,12 @@ func (r *MarketplaceOrderResource) Schema(ctx context.Context, req resource.Sche
 
 					float64planmodifier.UseStateForUnknown(),
 				}, MarkdownDescription: "Old Cost Estimate"},
+			"old_plan_billing_mode": schema.StringAttribute{
+				Computed: true,
+				PlanModifiers: []planmodifier.String{
+
+					stringplanmodifier.UseStateForUnknown(),
+				}, MarkdownDescription: "Old Plan Billing Mode"},
 			"old_plan_name": schema.StringAttribute{
 				Computed: true,
 				PlanModifiers: []planmodifier.String{